Move BaseHook imports to version_compat for standard provider
