+++ providers/standard/src/airflow/providers/standard/hooks/package_index.py
@@ -23,10 +23,7 @@ import subprocess
 from typing import Any
 from urllib.parse import quote, urlparse
 
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
+from airflow.providers.standard.version_compat import BaseHook
 
 
 class PackageIndexHook(BaseHook):
